src/requirements.txt
gitpython==3.1.40
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0